sysmon_eitl_analytic/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sysmon-eitl-analytic"
version = "0.1.0"
description = "Sysmon behavioral analytic redesigned as an Expert-in-the-Loop (EITL) system"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.0",
  "numpy>=1.24",
  "pyyaml>=6.0",
  "scikit-learn>=1.3",
  "pyarrow>=14.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4"]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["sysmon_pipeline*"]
